+++ tests/unit/base/test_commands_convert_onnx_onnx2trt.py
@@ -17,8 +17,10 @@ Note:
      Those test do not execute the conversion.
      The tests are checking if correct paths are executed on input arguments.
 """
+import json
 import pathlib
 import tempfile
+from unittest.mock import MagicMock
 
 import numpy as np
 
@@ -26,11 +28,12 @@ from model_navigator import TensorRTPrecision, TensorRTPrecisionMode
 from model_navigator.api.config import TensorRTProfile
 from model_navigator.commands.base import CommandStatus
 from model_navigator.commands.convert.onnx import ConvertONNX2TRT
+from model_navigator.execution_context import ExecutionContext
 from model_navigator.utils.tensor import TensorMetadata, TensorSpec
+from tests.utils import get_assets_path
 
 
 def test_run_execute_conversion_when_model_not_support_batching(mocker):
-
     with tempfile.TemporaryDirectory() as tmpdir:
         tmpdir = pathlib.Path(tmpdir)
         workspace = tmpdir / "navigator_workspace"
@@ -42,8 +45,8 @@ def test_run_execute_conversion_when_model_not_support_batching(mocker):
         output_model_path = workspace / "trt-fp16" / "model.plan"
         output_model_path.parent.mkdir(parents=True)
 
-        with mocker.patch.object(ConvertONNX2TRT, "_execute_conversion"), mocker.patch(
-            "model_navigator.commands.convert.onnx.onnx2trt.OnnxrtCPURunner"
+        with mocker.patch.object(ConvertONNX2TRT, "_execute_conversion"), mocker.patch.object(
+            ConvertONNX2TRT, "_get_onnx_input_metadata"
         ), mocker.patch("model_navigator.utils.devices.get_available_gpus", return_value=[0]):
             result = ConvertONNX2TRT().run(
                 workspace=workspace,
@@ -67,7 +70,6 @@ def test_run_execute_conversion_when_model_not_support_batching(mocker):
 
 
 def test_run_execute_conversion_when_trt_profile_provided(mocker):
-
     with tempfile.TemporaryDirectory() as tmpdir:
         tmpdir = pathlib.Path(tmpdir)
         workspace = tmpdir / "navigator_workspace"
@@ -79,8 +81,8 @@ def test_run_execute_conversion_when_trt_profile_provided(mocker):
         output_model_path = workspace / "trt-fp16" / "model.plan"
         output_model_path.parent.mkdir(parents=True)
 
-        with mocker.patch.object(ConvertONNX2TRT, "_execute_conversion"), mocker.patch(
-            "model_navigator.commands.convert.onnx.onnx2trt.OnnxrtCPURunner"
+        with mocker.patch.object(ConvertONNX2TRT, "_execute_conversion"), mocker.patch.object(
+            ConvertONNX2TRT, "_get_onnx_input_metadata"
         ), mocker.patch("model_navigator.utils.devices.get_available_gpus", return_value=[0]):
             result = ConvertONNX2TRT().run(
                 workspace=workspace,
@@ -105,7 +107,6 @@ def test_run_execute_conversion_when_trt_profile_provided(mocker):
 
 
 def test_run_execute_conversion_when_dataloader_and_device_max_batch_size_is_invalid(mocker):
-
     with tempfile.TemporaryDirectory() as tmpdir:
         tmpdir = pathlib.Path(tmpdir)
         workspace = tmpdir / "navigator_workspace"
@@ -117,8 +118,8 @@ def test_run_execute_conversion_when_dataloader_and_device_max_batch_size_is_inv
         output_model_path = workspace / "trt-fp16" / "model.plan"
         output_model_path.parent.mkdir(parents=True)
 
-        with mocker.patch.object(ConvertONNX2TRT, "_execute_conversion"), mocker.patch(
-            "model_navigator.commands.convert.onnx.onnx2trt.OnnxrtCPURunner"
+        with mocker.patch.object(ConvertONNX2TRT, "_execute_conversion"), mocker.patch.object(
+            ConvertONNX2TRT, "_get_onnx_input_metadata"
         ), mocker.patch("model_navigator.utils.devices.get_available_gpus", return_value=[0]):
             result = ConvertONNX2TRT().run(
                 workspace=workspace,
@@ -144,7 +145,6 @@ def test_run_execute_conversion_when_dataloader_and_device_max_batch_size_is_inv
 
 
 def test_run_execute_conversion_with_max_batch_size_search_when_dataloader_max_batch_size_provided(mocker):
-
     with tempfile.TemporaryDirectory() as tmpdir:
         tmpdir = pathlib.Path(tmpdir)
         workspace = tmpdir / "navigator_workspace"
@@ -156,9 +156,11 @@ def test_run_execute_conversion_with_max_batch_size_search_when_dataloader_max_b
         output_model_path = workspace / "trt-fp16" / "model.plan"
         output_model_path.parent.mkdir(parents=True)
 
-        with mocker.patch.object(ConvertONNX2TRT, "_execute_conversion_with_max_batch_size_search"), mocker.patch(
-            "model_navigator.commands.convert.onnx.onnx2trt.OnnxrtCPURunner"
-        ), mocker.patch("model_navigator.utils.devices.get_available_gpus", return_value=[0]):
+        with mocker.patch.object(
+            ConvertONNX2TRT, "_execute_conversion_with_max_batch_size_search"
+        ), mocker.patch.object(ConvertONNX2TRT, "_get_onnx_input_metadata"), mocker.patch(
+            "model_navigator.utils.devices.get_available_gpus", return_value=[0]
+        ):
             result = ConvertONNX2TRT().run(
                 workspace=workspace,
                 parent_path=input_model_path,
@@ -185,7 +187,6 @@ def test_run_execute_conversion_with_max_batch_size_search_when_dataloader_max_b
 
 
 def test_run_execute_conversion_with_max_batch_size_search_when_device_max_batch_size_provided(mocker):
-
     with tempfile.TemporaryDirectory() as tmpdir:
         tmpdir = pathlib.Path(tmpdir)
         workspace = tmpdir / "navigator_workspace"
@@ -197,9 +198,11 @@ def test_run_execute_conversion_with_max_batch_size_search_when_device_max_batch
         output_model_path = workspace / "trt-fp16" / "model.plan"
         output_model_path.parent.mkdir(parents=True)
 
-        with mocker.patch.object(ConvertONNX2TRT, "_execute_conversion_with_max_batch_size_search"), mocker.patch(
-            "model_navigator.commands.convert.onnx.onnx2trt.OnnxrtCPURunner"
-        ), mocker.patch("model_navigator.utils.devices.get_available_gpus", return_value=[0]):
+        with mocker.patch.object(
+            ConvertONNX2TRT, "_execute_conversion_with_max_batch_size_search"
+        ), mocker.patch.object(ConvertONNX2TRT, "_get_onnx_input_metadata"), mocker.patch(
+            "model_navigator.utils.devices.get_available_gpus", return_value=[0]
+        ):
             result = ConvertONNX2TRT().run(
                 workspace=workspace,
                 parent_path=input_model_path,
@@ -226,7 +229,6 @@ def test_run_execute_conversion_with_max_batch_size_search_when_device_max_batch
 
 
 def test_run_execute_conversion_with_max_batch_size_search_when_both_max_batch_size_provided(mocker):
-
     with tempfile.TemporaryDirectory() as tmpdir:
         tmpdir = pathlib.Path(tmpdir)
         workspace = tmpdir / "navigator_workspace"
@@ -238,9 +240,11 @@ def test_run_execute_conversion_with_max_batch_size_search_when_both_max_batch_s
         output_model_path = workspace / "trt-fp16" / "model.plan"
         output_model_path.parent.mkdir(parents=True)
 
-        with mocker.patch.object(ConvertONNX2TRT, "_execute_conversion_with_max_batch_size_search"), mocker.patch(
-            "model_navigator.commands.convert.onnx.onnx2trt.OnnxrtCPURunner"
-        ), mocker.patch("model_navigator.utils.devices.get_available_gpus", return_value=[0]):
+        with mocker.patch.object(
+            ConvertONNX2TRT, "_execute_conversion_with_max_batch_size_search"
+        ), mocker.patch.object(ConvertONNX2TRT, "_get_onnx_input_metadata"), mocker.patch(
+            "model_navigator.utils.devices.get_available_gpus", return_value=[0]
+        ):
             result = ConvertONNX2TRT().run(
                 workspace=workspace,
                 parent_path=input_model_path,
@@ -354,3 +358,69 @@ def test_get_shape_args_return_correct_arguments_when_batch_dim_and_max_batch_si
         "input_2:[128,256]",
     ]
     assert result == expected_result
+
+
+def test_get_onnx_input_metadata_return_filled_metadata_when_successfully_read_from_file(mocker):
+    with mocker.patch.object(
+        ExecutionContext, "execute_external_runtime_script"
+    ), tempfile.TemporaryDirectory() as tmpdir:
+        tmpdir = pathlib.Path(tmpdir)
+        results_file = tmpdir / "results.json"
+        workspace = tmpdir / "navigator_workspace"
+
+        mock = MagicMock()
+        mock.__enter__.return_value.name = results_file.as_posix()
+        mocker.patch("tempfile.NamedTemporaryFile", return_value=mock)
+
+        assets_path = get_assets_path()
+        model_path = assets_path / "models" / "identity.onnx"
+
+        input_metadata = TensorMetadata()
+        input_metadata.add(name="X", shape=(-1, 3, 8, 8), dtype=np.float32())
+
+        output_metadata = TensorMetadata()
+        output_metadata.add(name="Y", shape=(-1, 3, 8, 8), dtype=np.float32())
+
+        data = [{"name": "X", "shape": [-1, 3, -1, -1], "dtype": "float32"}]
+        with results_file.open("w") as fp:
+            json.dump(data, fp)
+
+        metadata = ConvertONNX2TRT()._get_onnx_input_metadata(
+            workspace=workspace,
+            input_model_path=model_path,
+            input_metadata=input_metadata,
+            output_metadata=output_metadata,
+            reproduce_script_path=workspace,
+            verbose=False,
+        )
+
+        assert "X" in metadata
+        assert metadata["X"] == TensorSpec(name="X", shape=(-1, 3, -1, -1), dtype=np.float32().dtype)
+
+
+def test_get_onnx_input_metadata_return_empty_metadata_when_no_file(mocker):
+    with mocker.patch.object(
+        ExecutionContext, "execute_external_runtime_script"
+    ), tempfile.TemporaryDirectory() as tmpdir:
+        tmpdir = pathlib.Path(tmpdir)
+        workspace = tmpdir / "navigator_workspace"
+
+        assets_path = get_assets_path()
+        model_path = assets_path / "models" / "identity.onnx"
+
+        input_metadata = TensorMetadata()
+        input_metadata.add(name="X", shape=(-1, 3, 8, 8), dtype=np.float32())
+
+        output_metadata = TensorMetadata()
+        output_metadata.add(name="Y", shape=(-1, 3, 8, 8), dtype=np.float32())
+
+        metadata = ConvertONNX2TRT()._get_onnx_input_metadata(
+            workspace=workspace,
+            input_model_path=model_path,
+            input_metadata=input_metadata,
+            output_metadata=output_metadata,
+            reproduce_script_path=workspace,
+            verbose=False,
+        )
+
+        assert metadata == {}